Misc refactoring
Also, persist user ID and device keys in shared preferences.
