+++ app/src/main/res/layout/activity_overview.xml
@@ -7,7 +7,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     android:paddingBottom="@dimen/activity_vertical_margin"
-    tools:context=".android.MainActivity">
+    tools:context=".DetailMainActivity">
     <HorizontalScrollView
         android:layout_width="fill_parent"
         android:layout_height="wrap_content">
